+++ CHANGELOG.md
@@ -39,6 +39,7 @@ Code](https://claude.com/product/claude-code).
 
 ## 1.0 Updated Key
 
++ Refator code for duplicate [1.0.15 20260717 gjw]
 + Allow DUPLICATE in entry view [1.0.14 20260717 gjw]
 + Updated oidc for chrome/web support [1.0.13 20260712 tonypioneer]
 + Update solidui/solidpod dependencies [1.0.12 20260703 gjw]